requirements.txt
Django==4.0.3
celery==5.2.3
Pillow==9.0.1
redis==4.1.4
django-tinymce-3.4.0